+++ ext/kernel/string.c
@@ -1,21 +1,13 @@
 /*
-  +------------------------------------------------------------------------+
-  | Zephir Language                                                        |
-  +------------------------------------------------------------------------+
-  | Copyright (c) 2011-2017 Zephir Team (http://www.zephir-lang.com)       |
-  +------------------------------------------------------------------------+
-  | This source file is subject to the New BSD License that is bundled     |
-  | with this package in the file docs/LICENSE.txt.                        |
-  |                                                                        |
-  | If you did not receive a copy of the license and are unable to         |
-  | obtain it through the world-wide-web, please send an email             |
-  | to license@zephir-lang.com so we can send you a copy immediately.      |
-  +------------------------------------------------------------------------+
-  | Authors: Andres Gutierrez <andres@zephir-lang.com>                     |
-  |          Eduar Carvajal <eduar@zephir-lang.com>                        |
-  |          Vladimir Kolesnikov <vladimir@extrememember.com>              |
-  +------------------------------------------------------------------------+
-*/
+ * This file is part of the Zephir.
+ *
+ * (c) Zephir Team <team@zephir-lang.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code. If you did not receive
+ * a copy of the license it is available through the world-wide-web at the
+ * following url: https://docs.zephir-lang.com/en/latest/license
+ */
 
 #ifdef HAVE_CONFIG_H
 #include "config.h"
@@ -1114,8 +1106,8 @@ void zephir_preg_match(zval *return_value, zval *regex, zval *subject, zval *mat
 	zval rv, tmp_matches;
 	zval *rvp = return_value ? return_value : &rv;
 
-	ZEPHIR_INIT_VAR(&tmp_flags);
-	ZEPHIR_INIT_VAR(&tmp_offset);
+	ZVAL_NULL(&tmp_flags);
+	ZVAL_NULL(&tmp_offset);
 
 	ZVAL_LONG(&tmp_flags, flags);
 	ZVAL_LONG(&tmp_offset, offset);
@@ -1191,7 +1183,7 @@ int zephir_json_encode(zval *return_value, zval *v, int opts)
 	zval zopts;
 	zval *params[2];
 
-	ZEPHIR_NULL(&zopts);
+	ZVAL_NULL(&zopts);
 	ZVAL_LONG(&zopts, opts);
 
 	params[0] = v;
@@ -1205,7 +1197,7 @@ int zephir_json_decode(zval *return_value, zval *v, zend_bool assoc)
 	zval zassoc;
 	zval *params[2];
 
-	ZEPHIR_NULL(&zassoc);
+	ZVAL_NULL(&zassoc);
 	ZVAL_BOOL(&zassoc, assoc);
 
 	params[0] = v;